+++ .eslintrc.js
@@ -259,7 +259,6 @@ module.exports = {
             {
                 allow: [
                     "**/*.css",
-                    "**/*.less",
                     "@testing-library/jest-dom",
                     "@testing-library/jest-dom/extend-expect",
                     "whatwg-fetch",