+++ packages/math-input/src/components/tabbar/tabbar.stories.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import Tabbar from "./tabbar";
 
 import type {KeypadPageType} from "../../types";
-import type {Meta, StoryObj} from "@storybook/react";
+import type {Meta, StoryObj} from "@storybook/react-vite";
 
 type Story = StoryObj<typeof Tabbar>;