+++ packages/perseus-core/src/parse-perseus-json/perseus-parsers/sorter-user-input.ts
@@ -1,0 +1,6 @@
+import {array, boolean, object, string} from "../general-purpose-parsers";
+
+export const parseSorterUserInput = object({
+    options: array(string),
+    changed: boolean,
+});