+++ packages/perseus-core/src/utils/split-perseus-item.ts
@@ -1,4 +1,5 @@
-import _ from "underscore";
+import {parseAndMigratePerseusItem} from "../parse-perseus-json";
+import {isFailure} from "../parse-perseus-json/result";
 
 import deepClone from "./deep-clone";
 import splitPerseusRenderer from "./split-perseus-renderer";
@@ -20,4 +21,23 @@ export default function splitPerseusItem(original: PerseusItem): PerseusItem {
         // so we consider that part of the answer data
         hints: [],
     };
+}
+
+/**
+ * Returns a JSON copy of a PerseusItem with rubric data (i.e. answers and
+ * hints) removed. Idempotent and deterministic.
+ *
+ * @param data a {@linkcode PerseusItem}, either as JSON or as an object.
+ * @returns {string} the answerless data formatted as JSON
+ * @throws {SyntaxError} given malformed JSON or data that can't be parsed as
+ * a Perseus item.
+ */
+export function splitPerseusItemJSON(data: unknown): string {
+    const parseResult = parseAndMigratePerseusItem(data);
+    if (isFailure(parseResult)) {
+        throw new SyntaxError(parseResult.detail.message);
+    }
+    const item = parseResult.value;
+
+    return JSON.stringify(splitPerseusItem(item));
 }