+++ packages/perseus-core/src/widgets/passage-ref/index.ts
@@ -9,6 +9,7 @@ const passageRefWidgetLogic: WidgetLogic = {
     version: currentVersion,
     defaultWidgetOptions: defaultWidgetOptions,
     defaultAlignment: "inline",
+    accessible: false,
 };
 
 export default passageRefWidgetLogic;