+++ packages/perseus/src/widgets/explanation/explanation.test.ts
@@ -198,7 +198,6 @@ describe("Explanation", function () {
     });
 
     it("does NOT use transitions when matchMedia is not available", async () => {
-        // e.g. during SSR in webapp
         const fakeWindow = Object.create(window);
         fakeWindow.matchMedia = undefined;
         jest.spyOn(window, "window", "get").mockReturnValue(fakeWindow as any);