+++ packages/perseus/src/widgets/radio/choice.new.tsx
@@ -18,7 +18,7 @@ import type {APIOptions} from "../../types";
 const intermediateCheckboxPadding = `16px 16px`;
 const intermediateCheckboxPaddingPhone = `12px 16px`;
 
-export type ChoiceProps = {
+export interface ChoiceProps {
     apiOptions?: APIOptions;
     checked?: boolean;
     rationale: React.ReactNode;
@@ -41,7 +41,7 @@ export type ChoiceProps = {
     // an object with a `checked` key. It contains a boolean value specifying
     // the new checked value of this choice.
     onChange?: (newValues: {checked: boolean}) => void;
-};
+}
 
 /**
  * This component is a duplicate of the Choice component in choice.tsx
@@ -114,7 +114,7 @@ const Choice = React.forwardRef<HTMLButtonElement, ChoiceProps>(
             <div className={descriptionClassName} style={styles.description}>
                 <div
                     style={{
-                        display: "flex",
+                        ...styles.choicesContainer,
                         opacity: showDimmed ? 0.5 : 1.0,
                     }}
                 >
@@ -143,11 +143,7 @@ const Choice = React.forwardRef<HTMLButtonElement, ChoiceProps>(
                             });
                         }}
                         disabled={disabled || apiOptions.readOnly}
-                        style={{
-                            flex: 1,
-                            color: color.offBlack,
-                            userSelect: "text",
-                        }}
+                        style={styles.optionContainer}
                         ref={ref as any}
                         aria-hidden="true"
                         hideDefaultFocusRing={true}
@@ -207,17 +203,26 @@ const styles = {
         color: color.offBlack,
     },
 
+    choicesContainer: {
+        display: "flex",
+    },
+
+    optionContainer: {
+        flex: 1,
+        color: color.offBlack,
+        userSelect: "text",
+    },
+
     choiceRow: {
         display: "flex",
         flexDirection: "row",
         justifyContent: "flex-start",
         alignItems: "center",
-        paddingTop: spacing.xSmall_8,
-        paddingBottom: spacing.xSmall_8,
-        paddingLeft: spacing.xSmall_8,
+        paddingBlock: spacing.xSmall_8,
+        paddingInlineStart: spacing.xSmall_8,
         position: "relative",
         minWidth: "100%",
-        left: 0,
+        insetInlineStart: 0,
         // The height and background will ensure content will be scrolled
         // behind the ChoiceIcon and will not be visible
         height: "100%",
@@ -229,10 +234,14 @@ const styles = {
         flexShrink: 0,
         flexBasis: "30px",
         position: "sticky",
-        left: 0,
+        insetInlineStart: 0,
         background: "linear-gradient(to right, white 70%, transparent)",
-        // TODO: LEMS-3107 adjust to support RTL
-        paddingRight: spacing.xSmall_8,
+        // TODO: LEMS-3107 not properly working fix this by moving in a css file
+        ":dir(rtl)": {
+            background: "linear-gradient(to left, white 70%, transparent)",
+        },
+        paddingInlineStart: spacing.xSmall_8,
+        paddingInlineEnd: spacing.xSmall_8,
         paddingTop: spacing.xxxSmall_4,
         zIndex: 1,
         alignItems: "center",
@@ -241,8 +250,8 @@ const styles = {
     },
 
     choiceContent: {
-        paddingLeft: spacing.small_12,
-        textAlign: "left",
+        paddingInlineStart: spacing.small_12,
+        textAlign: "start",
         flex: 1,
         paddingTop: spacing.xxxSmall_4,
         minWidth: 0,
@@ -252,7 +261,7 @@ const styles = {
         display: "block",
         padding: intermediateCheckboxPadding,
         paddingTop: 0,
-        marginLeft: 54,
+        marginInlineStart: 54,
         color: color.offBlack64,
         [mediaQueries.smOrSmaller]: {
             padding: intermediateCheckboxPaddingPhone,