+++ packages/perseus-core/CHANGELOG.md
@@ -1,5 +1,87 @@
 # @khanacademy/perseus-core
 
+## 14.0.0
+
+### Major Changes
+
+-   [#2485](https://github.com/Khan/perseus/pull/2485) [`cea05d9e6`](https://github.com/Khan/perseus/commit/cea05d9e6c629db6c14ed43dba085e215d006a1e) Thanks [@tatianasnook](https://github.com/tatianasnook)! - Migrate accessibility logic from perseus to perseus-core
+
+*   [#2474](https://github.com/Khan/perseus/pull/2474) [`59a2e87a8`](https://github.com/Khan/perseus/commit/59a2e87a83ef08a151552c7922bddfcbf1c73cf2) Thanks [@tatianasnook](https://github.com/tatianasnook)! - Moved checkAccessibilityAndWarn logic into a custom linter rule
+
+### Minor Changes
+
+-   [#2553](https://github.com/Khan/perseus/pull/2553) [`f5d08b79a`](https://github.com/Khan/perseus/commit/f5d08b79aca6d56ba7bdc3d47429c449645f75f9) Thanks [@catandthemachines](https://github.com/catandthemachines)! - Adding partically accessible widget function to interactive graph and label image.
+
+*   [#2532](https://github.com/Khan/perseus/pull/2532) [`75f1d1f53`](https://github.com/Khan/perseus/commit/75f1d1f53fe173f9275a213376c2497b5762ce0a) Thanks [@jeremywiebe](https://github.com/jeremywiebe)! - Add isItemAccessible helper which simplifies common usages of the violatingWidgets() function
+
+-   [#2518](https://github.com/Khan/perseus/pull/2518) [`0eba51c67`](https://github.com/Khan/perseus/commit/0eba51c67e1e67a11e425cb772194a44857eae0f) Thanks [@SonicScrewdriver](https://github.com/SonicScrewdriver)! - Updates to Radio Scoring to ensure non-specific multiselect questions are marked correctly.
+
+*   [#2526](https://github.com/Khan/perseus/pull/2526) [`ab2861d83`](https://github.com/Khan/perseus/commit/ab2861d8335e7ef1800dd5c6754616192b53073b) Thanks [@tatianasnook](https://github.com/tatianasnook)! - Pass data between EditorPage and ItemEditor via new props
+
+### Patch Changes
+
+-   [#2544](https://github.com/Khan/perseus/pull/2544) [`e60f6ac50`](https://github.com/Khan/perseus/commit/e60f6ac50a28d0664bfc211c9c655651719e3309) Thanks [@tatianasnook](https://github.com/tatianasnook)! - Move deleted traversal tests into perseus-core
+
+*   [#2545](https://github.com/Khan/perseus/pull/2545) [`cec57bbc6`](https://github.com/Khan/perseus/commit/cec57bbc6cf32cdce5a6b764b7399f69f87b4a89) Thanks [@Myranae](https://github.com/Myranae)! - Add recursive typesafe parser for Group's user input type with supporting utilities and widget ID parsing refactor
+
+## 13.0.0
+
+### Major Changes
+
+-   [#2419](https://github.com/Khan/perseus/pull/2419) [`843d66257`](https://github.com/Khan/perseus/commit/843d66257ceb32b7c3d3eea4dea47dfda58e7945) Thanks [@handeyeco](https://github.com/handeyeco)! - Remove the `propUpgrades` API from widgets as they're considered obsolete with the new parser
+
+### Patch Changes
+
+-   [#2549](https://github.com/Khan/perseus/pull/2549) [`404bde44d`](https://github.com/Khan/perseus/commit/404bde44dc6f8b6d5cbc8593c10b47f5f65d7eeb) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for the Free Response widget's user input type
+
+*   [#2541](https://github.com/Khan/perseus/pull/2541) [`d4720c470`](https://github.com/Khan/perseus/commit/d4720c4706888c922e3111c11f121a730d592aaf) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Grapher's user input type
+
+-   [#2550](https://github.com/Khan/perseus/pull/2550) [`35e4d13d9`](https://github.com/Khan/perseus/commit/35e4d13d9f6e49996e149950024a87f1601c2e43) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for CS Program's user input type
+
+*   [#2548](https://github.com/Khan/perseus/pull/2548) [`6bc9cc667`](https://github.com/Khan/perseus/commit/6bc9cc667c5bf2e48bfc5d036f11ba2285730839) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Iframe's user input type
+
+-   [#2479](https://github.com/Khan/perseus/pull/2479) [`074775b1f`](https://github.com/Khan/perseus/commit/074775b1ff0dc830a3d66030a14b776b0f631003) Thanks [@handeyeco](https://github.com/handeyeco)! - LEMS-2948: tweak how we determine when to wipe widget state clean (answerless/answerful)
+
+## 12.1.0
+
+### Minor Changes
+
+-   [#2533](https://github.com/Khan/perseus/pull/2533) [`03669d385`](https://github.com/Khan/perseus/commit/03669d385548e1d7ffde28c2210f637cf2b2d16f) Thanks [@benchristel](https://github.com/benchristel)! - Make `parseAndMigratePerseusItem` accept some legacy content items that previously caused parse errors.
+
+*   [#2495](https://github.com/Khan/perseus/pull/2495) [`86ac76f79`](https://github.com/Khan/perseus/commit/86ac76f79c2fd4ee8b822393f89a064af92c0339) Thanks [@Myranae](https://github.com/Myranae)! - Update label image widget to handle answerless data and do some cleanup
+
+-   [#2528](https://github.com/Khan/perseus/pull/2528) [`dc864aca8`](https://github.com/Khan/perseus/commit/dc864aca8ebf3bb442b1519f67d1a095d9f89243) Thanks [@benchristel](https://github.com/benchristel)! - Add `splitPerseusItemJSON` to `@khanacademy/perseus-core`, for use in the publish pipeline.
+
+### Patch Changes
+
+-   [#2538](https://github.com/Khan/perseus/pull/2538) [`52ec0bfcb`](https://github.com/Khan/perseus/commit/52ec0bfcbb71242c7a0a0ab34c7da0e9db77ed21) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Label-image's user input type
+
+*   [#2520](https://github.com/Khan/perseus/pull/2520) [`d1b6d9a3c`](https://github.com/Khan/perseus/commit/d1b6d9a3cab145267256e3d4acc373361957ffe0) Thanks [@Myranae](https://github.com/Myranae)! - Add parsing validation for Matrix's user input
+
+-   [#2523](https://github.com/Khan/perseus/pull/2523) [`6b747d140`](https://github.com/Khan/perseus/commit/6b747d1402cd62d9c59b2fc6f2371ec0a453a63d) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for the plotter user input type
+
+*   [#2516](https://github.com/Khan/perseus/pull/2516) [`efb99e901`](https://github.com/Khan/perseus/commit/efb99e9016418875a3b3713c8077a9465a5b41c6) Thanks [@Myranae](https://github.com/Myranae)! - Add parsing validation for Categorizer's user input
+
+-   [#2539](https://github.com/Khan/perseus/pull/2539) [`3ec658e38`](https://github.com/Khan/perseus/commit/3ec658e387d2f2193f8151a9d4cb285e85f13202) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Table's user input type
+
+*   [#2536](https://github.com/Khan/perseus/pull/2536) [`3e162e616`](https://github.com/Khan/perseus/commit/3e162e6168960ff1a1c63ec0e982422a113cbe31) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Orderer's user input type
+
+-   [#2521](https://github.com/Khan/perseus/pull/2521) [`0913e9397`](https://github.com/Khan/perseus/commit/0913e93973394270a474388f4c1e7c795027b015) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Number-line's user input type
+
+*   [#2524](https://github.com/Khan/perseus/pull/2524) [`8ece223ef`](https://github.com/Khan/perseus/commit/8ece223ef58ce23ccd244b6fb717193019e2c5d0) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Matcher's user input type
+
+-   [#2519](https://github.com/Khan/perseus/pull/2519) [`cf71982e0`](https://github.com/Khan/perseus/commit/cf71982e0fe9f831456d760fc4e98b1e93748c4f) Thanks [@dependabot](https://github.com/apps/dependabot)! - Updating wonderblock dependencies.
+
+*   [#2537](https://github.com/Khan/perseus/pull/2537) [`239be3f6c`](https://github.com/Khan/perseus/commit/239be3f6ccbbb89f52acf313bb4afbe359c1cee5) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Input-number's user input type
+
+-   [#2509](https://github.com/Khan/perseus/pull/2509) [`0d46a8599`](https://github.com/Khan/perseus/commit/0d46a85999557ca17daad67cb1663fe88a33a3ae) Thanks [@SonicScrewdriver](https://github.com/SonicScrewdriver)! - Ensure that Interactive Graph's labelLocation property can be parsed properly
+
+*   [#2525](https://github.com/Khan/perseus/pull/2525) [`1c49ad243`](https://github.com/Khan/perseus/commit/1c49ad243756b4519c57603c9d098cbfe18baa15) Thanks [@Myranae](https://github.com/Myranae)! - Add typesafe parser for Sorter's user input type
+
+*   Updated dependencies [[`cf71982e0`](https://github.com/Khan/perseus/commit/cf71982e0fe9f831456d760fc4e98b1e93748c4f)]:
+    -   @khanacademy/kas@2.0.5
+    -   @khanacademy/perseus-utils@2.0.4
+
 ## 12.0.0
 
 ### Major Changes